Extract tortls structures into a new header; clean up a little
